+++ test/e2e/specs/editor/various/inserting-blocks.spec.js
@@ -559,6 +559,57 @@ test.describe( 'Inserting blocks (@firefox, @webkit)', () => {
 			] );
 	} );
 
+	// Check for https://github.com/WordPress/gutenberg/issues/27540.
+	test( 'inserts a block before the first block using the inline inserter', async ( {
+		admin,
+		editor,
+		page,
+	} ) => {
+		await admin.createNewPost();
+		await editor.canvas
+			.getByRole( 'textbox', { name: 'Add title' } )
+			.fill( 'Post title' );
+		await editor.canvas
+			.getByRole( 'document', { name: 'Add default block' } )
+			.click();
+		await page.keyboard.type( 'First paragraph' );
+		await page.keyboard.press( 'Enter' );
+		await page.keyboard.type( 'Second paragraph' );
+
+		const boundingBox = await editor.canvas
+			.getByRole( 'document', { name: 'Block: Paragraph' } )
+			.filter( { hasText: 'First paragraph' } )
+			.boundingBox();
+
+		// Using the between inserter above the first block.
+		await page.mouse.move(
+			boundingBox.x + boundingBox.width / 2,
+			boundingBox.y - 10,
+			// An arbitrary number of `steps` imitates cursor movement in the test environment,
+			// activating the in-between inserter.
+			{ steps: 10 }
+		);
+
+		await page
+			.getByRole( 'button', {
+				name: 'Add block',
+			} )
+			.click();
+		await page.getByRole( 'button', { name: 'Browse All' } ).click();
+		await page
+			.getByRole( 'listbox', { name: 'Media' } )
+			.getByRole( 'option', { name: 'Image' } )
+			.click();
+
+		await expect
+			.poll( editor.getBlocks )
+			.toMatchObject( [
+				{ name: 'core/image' },
+				{ name: 'core/paragraph' },
+				{ name: 'core/paragraph' },
+			] );
+	} );
+
 	// Check for regression of https://github.com/WordPress/gutenberg/issues/72297.
 	test( 'keeps the inline inserter open when the Block Library panel closes', async ( {
 		admin,